refactor(transcribe): convert all transcribe resources to libnuke resource format
